move get_all_ops_desc from check_op_desc.py to print_op_desc.py
