+++ tools/check_op_desc.py
@@ -12,10 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import paddle.fluid.framework as framework
-from paddle.fluid import core
 import json
-from paddle import compat as cpt
 import sys
 
 SAME = 0
@@ -37,52 +34,6 @@ GENERATED = "generated"
 DEFAULT_VALUE = "default_value"
 
 error = False
-
-
-def get_attr_default_value(op_name):
-    return core.get_op_attrs_default_value(cpt.to_bytes(op_name))
-
-
-def get_vars_info(op_vars_proto):
-    vars_info = {}
-    for var_proto in op_vars_proto:
-        name = str(var_proto.name)
-        vars_info[name] = {}
-        vars_info[name][DUPLICABLE] = var_proto.duplicable
-        vars_info[name][DISPENSABLE] = var_proto.dispensable
-        vars_info[name][INTERMEDIATE] = var_proto.intermediate
-    return vars_info
-
-
-def get_attrs_info(op_proto, op_attrs_proto):
-    attrs_info = {}
-    attrs_default_values = get_attr_default_value(op_proto.type)
-    for attr_proto in op_attrs_proto:
-        attr_name = str(attr_proto.name)
-        attrs_info[attr_name] = {}
-        attrs_info[attr_name][TYPE] = attr_proto.type
-        attrs_info[attr_name][GENERATED] = attr_proto.generated
-        attrs_info[attr_name][DEFAULT_VALUE] = attrs_default_values[
-            attr_name] if attr_name in attrs_default_values else None
-    return attrs_info
-
-
-def get_op_desc(op_proto):
-    op_info = {}
-    op_info[INPUTS] = get_vars_info(op_proto.inputs)
-    op_info[OUTPUTS] = get_vars_info(op_proto.outputs)
-    op_info[ATTRS] = get_attrs_info(op_proto, op_proto.attrs)
-    return op_info
-
-
-def get_all_ops_desc():
-    all_op_protos_dict = {}
-    all_op_protos = framework.get_all_op_protos()
-    for op_proto in all_op_protos:
-        op_type = str(op_proto.type)
-        all_op_protos_dict[op_type] = get_op_desc(op_proto)
-
-    return all_op_protos_dict
 
 
 def diff_vars(origin_vars, new_vars):
@@ -268,34 +219,7 @@ def print_error_message(error_message):
                     format(arg, name, ori_value, new_value))
 
 
-if len(sys.argv) == 1:
-    '''
-    Print all ops desc in dict:
-    {op1_name:
-        {INPUTS:
-            {input_name1:
-                {DISPENSABLE: bool,
-                 INTERMEDIATE: bool,
-                 DUPLICABLE: bool
-                },
-            input_name2:{}    
-            },
-         OUTPUTS:{},
-         ATTRS:
-            {attr_name1:
-                {TYPE: int,
-                 GENERATED: bool,
-                 DEFAULT_VALUE: int/str/etc,
-                }
-            }    
-        }
-     op2_name:{}        
-    }
-    '''
-    all_op_protos_dict = get_all_ops_desc()
-    result = json.dumps(all_op_protos_dict)
-    print(result)
-elif len(sys.argv) == 3:
+if len(sys.argv) == 3:
     '''
     Compare op_desc files generated by branch DEV and branch PR.
     And print error message.
@@ -311,7 +235,4 @@ elif len(sys.argv) == 3:
         print_error_message(error_message)
 
 else:
-    print("Usage:\n" \
-          "\t1. python check_op_desc.py > OP_DESC_DEV.spec\n" \
-          "\t2. python check_op_desc.py > OP_DESC_PR.spec\n"\
-          "\t3. python check_op_desc.py OP_DESC_DEV.spec OP_DESC_PR.spec > error_message")
+    print("Usage: python check_op_desc.py OP_DESC_DEV.spec OP_DESC_PR.spec")
